refactor: alter id type for roles, courses, and associated tables from bigint to int

+++ src/main/resources/db/migration/V3__create_users_roles_table.sql
@@ -1,6 +1,6 @@
 CREATE TABLE users_roles (
     user_id BIGINT NOT NULL,
-    role_id BIGINT NOT NULL,
+    role_id INT NOT NULL,
     PRIMARY KEY (user_id, role_id),
     FOREIGN KEY (user_id) REFERENCES users (id),
     FOREIGN KEY (role_id) REFERENCES roles (id)
